refactor: extract ProdUplcConstrEmitter (Phase 4c step 3b)
Move the native-UplcConstr Constr emission method out of
ProductCaseSirTypeGenerator into a new ProdUplcConstrEmitter:

  - genConstr (was genConstrUplcConstr)
  - private hasClassLevelUplcRepr helper

Updates the five cross-class call sites to call
ProdUplcConstrEmitter.genConstr directly:

  - ProductCaseSirTypeGenerator.genConstrLowered (2x)
  - ProductCaseUplcOnlySirTypeGenerator
  - ProductCaseUplcConstrSirTypeGenerator
  - SumCaseUplcConstrSirTypeGenerator
  - SumCaseUplcOnlySirTypeGenerator

retrieveConstrIndex / retrieveConstrDecl stay on
ProductCaseSirTypeGenerator and are referenced cross-class from the
new emitter.
