+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumCaseUplcOnlySirTypeGenerator.scala
@@ -42,7 +42,7 @@ object SumCaseUplcOnlySirTypeGenerator extends SirTypeUplcGenerator {
         loweredArgs: scala.List[LoweredValue],
         optTargetType: Option[SIRType]
     )(using lctx: LoweringContext): LoweredValue = {
-        ProductCaseSirTypeGenerator.genConstrUplcConstr(constr, loweredArgs)
+        ProdUplcConstrEmitter.genConstr(constr, loweredArgs)
     }
 
     override def genSelect(sel: SIR.Select, loweredScrutinee: LoweredValue)(using
